requirements.txt
c2pa-python==0.4.0
PyExifTool==0.5.6
python-dotenv==1.0.1
